Auth0 infra

+++ backend/src/api/coaches.ts
@@ -1,6 +1,6 @@
 import { Router } from "express"
 import { Error } from "mongoose"
-import { ensureAdmin, ensureLogin, getIdentity } from "../auth"
+import { createUser, deleteUser, ensureAdmin, ensureLogin, getIdentity, updateUser } from "../auth"
 import { CoachModel, type Coach } from "../schemas/coach"
 
 const router = Router()
@@ -14,6 +14,10 @@ router.get("/", async (_req, res) => {
 })
 
 router.get("/:id", async (req, res, next) => {
+  if (!req.params.id) {
+    return res.status(404).json({ error: "Not found" })
+  }
+
   try {
     const item = await CoachModel.findById(req.params.id)
 
@@ -41,6 +45,13 @@ router.post("/", ensureAdmin, async (req, res, next) => {
 
   try {
     await item.save()
+    const { error } = await createUser(item)
+
+    if (error) {
+      await CoachModel.deleteOne({ _id: item._id })
+      return next(error)
+    }
+
     return res.status(201).json(item)
   } catch (e) {
     if (e instanceof Error.ValidationError) {
@@ -59,12 +70,29 @@ router.patch("/:id", ensureAdmin, async (req, res, next) => {
       return res.status(404).json({ error: "Not found" })
     }
 
+    const oldItem = {
+      name: item.name,
+      email: item.email,
+      admin: item.admin,
+    }
+
     const { name, email, admin } = req.body as Partial<Coach>
     item.name = name ?? item.name
     item.email = email ?? item.email
     item.admin = admin ?? item.admin
 
     await item.save()
+    const { error } = await updateUser(item)
+
+    if (error) {
+      item.name = oldItem.name
+      item.email = oldItem.email
+      item.admin = oldItem.admin
+
+      await item.save()
+      return next(error)
+    }
+
     return res.status(201).json(item)
   } catch (e) {
     if (e instanceof Error.DocumentNotFoundError) {
@@ -87,12 +115,29 @@ router.put("/:id", ensureAdmin, async (req, res, next) => {
       return res.status(404).json({ error: "Not found" })
     }
 
+    const oldItem = {
+      name: item.name,
+      email: item.email,
+      admin: item.admin,
+    }
+
     const { name, email, admin } = req.body as Coach
     item.name = name
     item.email = email
-    item.admin = admin
+    item.admin = admin ?? false
 
     await item.save()
+    const { error } = await updateUser(item)
+
+    if (error) {
+      item.name = oldItem.name
+      item.email = oldItem.email
+      item.admin = oldItem.admin
+
+      await item.save()
+      return next(error)
+    }
+
     return res.status(201).json(item)
   } catch (e) {
     if (e instanceof Error.DocumentNotFoundError) {
@@ -108,8 +153,25 @@ router.put("/:id", ensureAdmin, async (req, res, next) => {
 })
 
 router.delete("/:id", ensureAdmin, async (req, res, next) => {
+  if (!req.params.id) {
+    return res.status(404).json({ error: "Not found" })
+  }
+
   try {
+    const oldItem = await CoachModel.findById(req.params.id)
+
+    if (!oldItem) {
+      return res.status(404).json({ error: "Not found" })
+    }
+
     await CoachModel.deleteOne({ _id: req.params.id })
+    const { error } = deleteUser(req.params.id)
+
+    if (error) {
+      await oldItem.save()
+      return next(error)
+    }
+
     return res.status(204).json()
   } catch (e) {
     if (e instanceof Error.DocumentNotFoundError) {
